Fix argument processing
This commit fixes the argument processing for this script.

1. Remove the if...elseif... block because this is notoriously slow
2. When passing only one argument and it's not "--cached", instead of
doing nothing at all, actually show that an "Invalid argument" was passed
and display the usage text
3. This script never set any variables if more than 2 arguments were
passed. I'm assuming more than two arguments is an error, and so the catch
all case statement will apply showing the help usage text.
4. In the case of no arguments being passed, show the help usage text and
exit.
5. Don't create a temp file if invalid arguments are passed.
6. Create the temp file in the directory where this script is invoked
from, not in the directory where the script resides.